+++ client/src/pages/HomePage.jsx
@@ -1,7 +1,9 @@
 import { getCurrentSeason } from '../utils/constants'
 import { useSeasonalAnime } from '../hooks/useAnime'
 import HeroCarousel from '../components/anime/HeroCarousel'
+import TrendingSection from '../components/home/TrendingSection'
 import ContinueWatching from '../components/anime/ContinueWatching'
+import ActivityFeed from '../components/social/ActivityFeed'
 import WeeklySchedule from '../components/anime/WeeklySchedule'
 import LoadingSpinner from '../components/common/LoadingSpinner'
 
@@ -24,9 +26,11 @@ export default function HomePage() {
         : <HeroCarousel animeList={top5} />
       }
 
-      {/* Continue watching + Weekly schedule */}
+      {/* Trending + Continue watching + Weekly schedule */}
       <div className="container" style={{ paddingTop: 8, paddingBottom: 60 }}>
+        <TrendingSection />
         <ContinueWatching />
+        <ActivityFeed />
         <WeeklySchedule />
       </div>
     </div>